#10 EampleBlockReduce.java ausführbar

+++ examples/kernels/block_reduce.cu
@@ -1,8 +1,8 @@
-#import "block_reduce.h"
+#include "block_reduce.h"
 
 extern "C" __global__
 void deviceReduceKernel(long* in, long* out, int N) {
-    long long sum = 0;
+    long sum = 0;
 
     //reduce multiple elements per thread
     for (int i = blockIdx.x * blockDim.x + threadIdx.x; i < N; i += blockDim.x * gridDim.x) {
