+++ examples/kernels/warp_reduce.h
@@ -1,7 +1,7 @@
 #pragma once
 
 __inline__ __device__
-long long warpReduceSum(long long val) {
+long warpReduceSum(long val) {
 	for (int offset = warpSize / 2; offset > 0; offset /= 2)
 		val += __shfl_down_sync(0xFFFFFFFF, val, offset);
 	return val;
